\name{RcppCopyExample-package}
\alias{RcppCopyExample-package}
\alias{RcppCopyExample}
\docType{package}
\title{\packageTitle{RcppCopyExample}}
\description{\packageDescription{RcppCopyExample}}
\section{Package Content}{\packageIndices{RcppCopyExample}}
\author{\packageAuthor{RcppCopyExample}}
\section{Maintainer}{\packageMaintainer{RcppCopyExample}}
\keyword{package}
